Refactoring
Refactored numerals_list <-> shorthand_list conversions into sidewinder utilities
