+++ tests/test_basic.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 # Note: using `pip install -e .` as per https://docs.pytest.org/en/latest/explanation/goodpractices.html#goodpractices
 
-#from .context import sidewinder
 import sidewinder
 
 import unittest
@@ -19,8 +18,9 @@ class BasicTestSuite(unittest.TestCase):
         assert synonyms is not None
 
 
-class CoreExamplesRefactoring(unittest.TestCase):
-    """Tests to drive work on refactor-core 28/03/21 (creating a Chart objective to replicate monolithic examples of chords_to_midi(), chords_to_bassline_midi())"""
+class RepresentationConversions(unittest.TestCase):
+    """Tests to drive work on refactor-core 28/03/21 (creating a Chart object to replicate monolithic examples of chords_to_midi(), chords_to_bassline_midi())
+    Specifically conversion between the various representations of a Chart (numerals, shorthands)"""
 
     misty_numerals = 'IM7, v-7, I7, IVM7, iv-9, bVII7, IM7, vi-7, ii-7, V7, iii-7, VI7, ii-7, V7, \
     IM7, v-7, I7, IVM7, iv-9, bVII7, IM7, vi-7, ii-7, V7, I6, bVII9, IM7, \
@@ -84,6 +84,24 @@ class CoreExamplesRefactoring(unittest.TestCase):
         # most likely using the sidewinder.detect_numeral_pattern() function which should be refactored similarly with utilities.progression_to_chords() / Chart.numerals_list_to_shorthand_list()
 
         assert False
+
+class AddingDurationsToProgressions(unittest.TestCase):
+    """Tests to drive work on refactor-core 28/03/21 (creating a Chart object to replicate monolithic examples of chords_to_midi(), chords_to_bassline_midi())
+    Specifically overlaying durations to the Chart representations to create compositions e.g. voicings -> midi files"""
+
+    misty_numerals = 'IM7, v-7, I7, IVM7, iv-9, bVII7, IM7, vi-7, ii-7, V7, iii-7, VI7, ii-7, V7, \
+    IM7, v-7, I7, IVM7, iv-9, bVII7, IM7, vi-7, ii-7, V7, I6, bVII9, IM7, \
+    v-7, I7b9, IVM7, IVM7,\
+    bv-7, VII7, II7, iii-7, VI7b9, ii-7, V7, \
+    IM7, v-7, I7, IVM7, iv-9, bVII7, IM7, vi-7, ii-7, V7, I6, I6'
+
+    misty_key = 'F'
+
+    misty_durs = [1, 2, 2, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 
+            1, 2, 2, 1, 2, 2, 2, 2, 2, 2, 2, 2, 1, 
+            1, 1, 1, 1, 
+            1, 2, 2, 2, 2, 2, 2, 
+            1, 2, 2, 1, 2, 2, 2, 2, 2, 2, 1, 1]
 
     def test_create_representation_of_chart_with_durations(self):
